chore: extract auth-client into auth-client/ folder for migration boundary
