+++ src/server/auth-client/index.ts
@@ -1,0 +1,3 @@
+export * from "./auth-client.js";
+export * from "./connect-account-errors.js";
+export * from "./types.js";
