+++ src/server/mfa/mfa-error-bubbling.test.ts
@@ -20,7 +20,7 @@ import {
 import { createNextHeadersMock } from "../../test-fixtures/mocks.js";
 import { generateSecret } from "../../test-fixtures/utils.js";
 import { MfaContext, SessionData } from "../../types/index.js";
-import { AuthClient } from "../auth-client.js";
+import { AuthClient } from "../auth-client/index.js";
 import { Auth0Client } from "../client.js";
 import { decrypt } from "../cookies/index.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
