+++ src/server/passwordless/passwordless-server.flow.test.ts
@@ -13,7 +13,7 @@ import {
 import { generateSecret, stripTransactionValuePrefix } from "../../test-fixtures/utils.js";
 import type { SessionData } from "../../types/index.js";
 import { AuthClientProvider } from "../auth-client-provider.js";
-import { AuthClient } from "../auth-client.js";
+import { AuthClient } from "../auth-client/index.js";
 import { decrypt, encrypt } from "../cookies/index.js";
 import { generateDpopKeyPair } from "../dpop/retry.js";
 import { ServerPasswordlessClient } from "../passwordless/server-passwordless-client.js";
