+++ src/server/passwordless/server-passwordless-client.test.ts
@@ -19,7 +19,7 @@ import {
   setupMswLifecycle
 } from "../../test-fixtures/defaults.js";
 import { AuthClientProvider } from "../auth-client-provider.js";
-import { AuthClient } from "../auth-client.js";
+import { AuthClient } from "../auth-client/index.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
 import { TransactionStore } from "../transaction-store.js";
 import { ServerPasswordlessClient } from "./server-passwordless-client.js";
