+++ src/server/passwordless/server-passwordless-client.ts
@@ -11,7 +11,7 @@ import type {
   PasswordlessVerifyOptions
 } from "../../types/index.js";
 import type { AuthClientProvider } from "../auth-client-provider.js";
-import type { AuthClient } from "../auth-client.js";
+import type { AuthClient } from "../auth-client/index.js";
 
 /**
  * Server-side passwordless authentication API.
