+++ src/server/tests/auth-client-provider.test.ts
@@ -7,7 +7,7 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { InvalidConfigurationError } from "../../errors/index.js";
 import { DomainResolutionError } from "../../errors/mcd.js";
 import { AuthClientProvider } from "../auth-client-provider.js";
-import type { AuthClient } from "../auth-client.js";
+import type { AuthClient } from "../auth-client/index.js";
 
 describe("AuthClientProvider", () => {
   let mockAuthClient: AuthClient;
