+++ src/server/tests/redundant-txn-cookie-deletion.test.ts
@@ -18,7 +18,7 @@ import { InvalidStateError, MissingStateError } from "../../errors/index.js";
 import { getDefaultRoutes } from "../../test-fixtures/defaults.js";
 import { generateSecret } from "../../test-fixtures/utils.js";
 import { SessionData } from "../../types/index.js";
-import { AuthClient, AuthClientOptions } from "../auth-client.js";
+import { AuthClient, AuthClientOptions } from "../auth-client/index.js";
 import {
   ReadonlyRequestCookies,
   RequestCookies,
